improve R measurements processing
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Addons/EventSim/trunk@31242 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ edu.kit.ipd.sdq.eventsim.measurement.r/src/edu/kit/ipd/sdq/eventsim/measurement/r/RStatistics.java
@@ -2,13 +2,13 @@ package edu.kit.ipd.sdq.eventsim.measurement.r;
 
 public class RStatistics {
 
-	private double timeSpentInR; // in milliseconds
+	private long timeSpentInR; // in milliseconds
 
-	public void captureTimeSpentInR(double millisecondsSpentInR) {
+	public void captureTimeSpentInR(long millisecondsSpentInR) {
 		timeSpentInR += millisecondsSpentInR;
 	}
 
-	public double getTotalTimeSpentInR() {
+	public long getTotalTimeSpentInR() {
 		return timeSpentInR;
 	}
